ll::Lbl is now a u32 + StrInterner type for reuse
onto the next! Uids will be very difficult, as they should be assigned
in the parser, but the parse ctx is Copy!

+++ src/backend/x64/ast.rs
@@ -1,5 +1,5 @@
 // index into Prog.labels
-pub type Label = usize;
+pub type Label = u32;
 
 #[derive(Clone, Copy, Debug)]
 pub enum Reg {
